+++ packages/contracts/hardhat.config.ts
@@ -5,9 +5,12 @@ import "hardhat-gas-reporter";
 import "hardhat-deploy";
 import "solidity-coverage";
 import "@tenderly/hardhat-tenderly";
+import "@nomiclabs/hardhat-etherscan";
 
 import { config as dotEnvConfig } from "dotenv";
 import { HardhatUserConfig } from "hardhat/types";
+import { TASK_COMPILE_SOLIDITY_GET_SOLC_BUILD } from "hardhat/builtin-tasks/task-names";
+import { subtask } from "hardhat/config";
 
 import "./src/tasks/addRouter";
 import "./src/tasks/addAsset";
@@ -34,6 +37,26 @@ const mnemonic =
   process.env.SUGAR_DADDY ||
   process.env.MNEMONIC ||
   "candy maple cake sugar pudding cream honey rich smooth crumble sweet treat";
+
+const path = require("path");
+
+subtask(TASK_COMPILE_SOLIDITY_GET_SOLC_BUILD, async (args: any, hre, runSuper) => {
+  if (args.solcVersion === "0.8.4") {
+    const compilerPath = path.join(__dirname, "soljson.js");
+
+    return {
+      compilerPath,
+      isSolcJs: true, // if you are using a native compiler, set this to false
+      version: args.solcVersion,
+      // this is used as extra information in the build-info files, but other than
+      // that is not important
+      longVersion: "0.8.4",
+    };
+  }
+
+  // we just use the default subtask if the version is not 0.8.5
+  return runSuper();
+});
 
 const config: HardhatUserConfig = {
   solidity: {
@@ -178,6 +201,9 @@ const config: HardhatUserConfig = {
       chainId: 421611,
       url: urlOverride || process.env.ARB_RINK_ETH_PROVIDER_URL || "https://rinkeby.arbitrum.io/rpc",
     },
+  },
+  etherscan: {
+    apiKey: process.env.ETHERSCAN_API_KEY,
   },
 };