+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Next Release
 
 - [contracts] Add ability to one-off verify contracts via hardhat
+- [router] Sanity check for router contracts existing on chain
+- [utils] Correctly determine the Gelato-supported chains
+- [router][txservice] Move sanity check of oracle address to txservice
 
 ## 0.1.10